requirements.txt
django-debug-toolbar
numpy